Bugfixes
- removed <emmintrin.h from platform code for GOST-3411
- added missed conditions for __int128 in dtoa code
- added correct enable_if for std::basic_ostream<char16_t> (it available
only when std::ctype<char16_t> defined
- fixed link-time deps for libzip on android

+++ core/utils/SPMovingAverage.h
@@ -28,7 +28,7 @@ THE SOFTWARE.
 
 namespace stappler::math {
 
-template <size_t Count, typename T = float>
+template <uint64_t Count, typename T = float>
 class MovingAverage {
 public:
 	void dropValues() {
